+++ src/instagram.nim
@@ -1,6 +1,16 @@
 import instagram/core except get, endpoint
 export core
 
+# Types
+import instagram/api/types/user
+export user except IgUserResponse
+import instagram/api/types/followersAndFollowing
+export followersAndFollowing
+import instagram/api/types/post
+export post
+import instagram/api/types/feed
+export feed
+
 # API
 import instagram/api/get
 export get