+++ src/instagram/api/post.nim
@@ -39,3 +39,29 @@ proc unfollow*(
                                       body)
   if json{"friendship_status", "following"}.getBool:
     raise newException(IgException, "Cannot unfollow user '" & userId & "'")
+
+import instagram/api/types/post
+
+proc like*(
+  ig: Instagram;
+  postId: string or IgPost
+) {.async.} =
+  ## Likes the provided post
+  ## Error will raise an exception
+  when postId is IgPost:
+    let postId = postId.id
+  let json = parseJson await ig.request(HttpPost, endpoint("web/likes/$#/like/", postId))
+  if json{"status"}.getStr != "ok":
+    raise newException(IgException, "Cannot like post '" & postId & "'")
+
+proc unlike*(
+  ig: Instagram;
+  postId: string or IgPost
+) {.async.} =
+  ## Likes the provided post
+  ## Error will raise an exception
+  when postId is IgPost:
+    let postId = postId.id
+  let json = parseJson await ig.request(HttpPost, endpoint("web/likes/$#/unlike/", postId))
+  if json{"status"}.getStr != "ok":
+    raise newException(IgException, "Cannot like post '" & postId & "'")